app/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
user-agents==2.2.0
psycopg2-binary==2.9.7